fix(@ngtools/webpack): report diagnostics the same way in type checker
