+++ packages/ngtools/webpack/src/angular_compiler_plugin.ts
@@ -28,7 +28,6 @@ import {
   createCompilerHost,
   createProgram,
   formatDiagnostics,
-  isNgDiagnostic,
   readConfiguration,
 } from '@angular/compiler-cli';
 import { ChildProcess, ForkOptions, fork } from 'child_process';
@@ -38,8 +37,8 @@ import * as ts from 'typescript';
 import { Compiler, compilation } from 'webpack';
 import { time, timeEnd } from './benchmark';
 import { WebpackCompilerHost } from './compiler_host';
+import { DiagnosticMode, gatherDiagnostics, hasErrors, reportDiagnostics } from './diagnostics';
 import { resolveEntryModuleFromMain } from './entry_resolver';
-import { DiagnosticMode, gatherDiagnostics, hasErrors } from './gather_diagnostics';
 import {
   AngularCompilerPluginOptions,
   ContextElementDependencyConstructor,
@@ -1097,59 +1096,13 @@ export class AngularCompilerPlugin {
     const { emitResult, diagnostics } = this._emit();
     timeEnd('AngularCompilerPlugin._update._emit');
 
-    // Report Diagnostics
-    const tsErrors = [];
-    const tsWarnings = [];
-    const ngErrors = [];
-    const ngWarnings = [];
-
-    for (const diagnostic of diagnostics) {
-      switch (diagnostic.category) {
-        case ts.DiagnosticCategory.Error:
-          if (isNgDiagnostic(diagnostic)) {
-            ngErrors.push(diagnostic);
-          } else {
-            tsErrors.push(diagnostic);
-          }
-          break;
-        case ts.DiagnosticCategory.Message:
-        case ts.DiagnosticCategory.Suggestion:
-          // Warnings?
-        case ts.DiagnosticCategory.Warning:
-          if (isNgDiagnostic(diagnostic)) {
-            ngWarnings.push(diagnostic);
-          } else {
-            tsWarnings.push(diagnostic);
-          }
-          break;
-      }
-    }
-
-    if (tsErrors.length > 0) {
-      const message = ts.formatDiagnosticsWithColorAndContext(
-        tsErrors,
-        this._compilerHost,
-      );
-      this._errors.push(new Error(message));
-    }
-
-    if (tsWarnings.length > 0) {
-      const message = ts.formatDiagnosticsWithColorAndContext(
-        tsWarnings,
-        this._compilerHost,
-      );
-      this._warnings.push(message);
-    }
-
-    if (ngErrors.length > 0) {
-      const message = formatDiagnostics(ngErrors);
-      this._errors.push(new Error(message));
-    }
-
-    if (ngWarnings.length > 0) {
-      const message = formatDiagnostics(ngWarnings);
-      this._warnings.push(message);
-    }
+    // Report any diagnostics.
+    reportDiagnostics(
+      diagnostics,
+      this._compilerHost,
+      msg => this._errors.push(new Error(msg)),
+      msg => this._warnings.push(msg),
+    );
 
     this._emitSkipped = !emitResult || emitResult.emitSkipped;
